move colored label methods from FxTexts to FxLabels
